更新car_config.h和main.h中的测试模式配置，调整resetMode_start_time的数据类型为u32，以支持更大的时间值

+++ Core/Inc/main.h
@@ -73,7 +73,7 @@ extern int Task3_Move_Forward_Speed;     // 任务3中直线行驶的速度(速度值)
 extern u16 forwardBase_PWM;              // 直线行走基础PWM值
 extern u16 openLoopSteeringBase_PWM;     // 转向基础PWM值
 extern u8 resetTask_flag;                // 是否要重启任务标志位   
-extern u16 resetMode_start_time;         // 任务3重置模式开始时间
+extern u32 resetMode_start_time;         // 任务3重置模式开始时间
 /* USER CODE END ET */
 
 /* Exported constants --------------------------------------------------------*/
